# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/accounts": {
      "post": {
        "description": "Create a new financial account (bank, card, cash, etc.). Subject to plan limits.",
        "operationId": "AccountsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or invalid credit card fields"
          },
          "403": {
            "description": "Account limit reached for current plan"
          },
          "409": {
            "description": "Account name already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create account",
        "tags": [
          "Accounts"
        ]
      },
      "get": {
        "description": "Get all accounts for the authenticated workspace. By default, archived accounts are hidden.",
        "operationId": "AccountsController_findAll_v1",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by account type",
            "schema": {
              "example": "DEBIT_CARD",
              "type": "string",
              "enum": [
                "CASH",
                "DEBIT_CARD",
                "CREDIT_CARD",
                "SAVINGS",
                "INVESTMENT",
                "DIGITAL_WALLET",
                "OTHER"
              ]
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "description": "Include archived accounts (default: false)",
            "schema": {
              "default": false,
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "currencyCode",
            "required": false,
            "in": "query",
            "description": "Filter by currency code",
            "schema": {
              "example": "USD",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of accounts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List accounts",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/summary": {
      "get": {
        "description": "Get total balance grouped by currency across all active accounts that are included in total.",
        "operationId": "AccountsController_getBalanceSummary_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Balance summary by currency",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSummaryResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Balance summary",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/{id}": {
      "get": {
        "description": "Get details of a specific account.",
        "operationId": "AccountsController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get account",
        "tags": [
          "Accounts"
        ]
      },
      "put": {
        "description": "Update account details. Type and currency cannot be changed after creation.",
        "operationId": "AccountsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Account name already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update account",
        "tags": [
          "Accounts"
        ]
      },
      "delete": {
        "description": "Permanently delete an account. Only possible if the account has no transactions. Use archive instead for accounts with transaction history.",
        "operationId": "AccountsController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Account deleted successfully"
          },
          "404": {
            "description": "Account not found"
          },
          "409": {
            "description": "Account has transactions, archive instead"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete account",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/{id}/archive": {
      "patch": {
        "description": "Archive an account to hide it from the UI while preserving transaction history. Archived accounts do not count towards plan limits.",
        "operationId": "AccountsController_archive_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account archived successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Archive account",
        "tags": [
          "Accounts"
        ]
      }
    },
    "/api/v1/accounts/{id}/unarchive": {
      "patch": {
        "description": "Restore an archived account. Subject to plan limits — unarchiving may fail if you are at your account limit.",
        "operationId": "AccountsController_unarchive_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Account ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account unarchived successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Account limit reached, cannot unarchive"
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Unarchive account",
        "tags": [
          "Accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Facets API",
    "description": "Professional Finance Tracker SaaS API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Account name",
            "example": "My Checking Account",
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "description": "Account type",
            "enum": [
              "CASH",
              "DEBIT_CARD",
              "CREDIT_CARD",
              "SAVINGS",
              "INVESTMENT",
              "DIGITAL_WALLET",
              "OTHER"
            ],
            "example": "DEBIT_CARD"
          },
          "balance": {
            "type": "number",
            "description": "Initial balance (defaults to 0)",
            "example": 1500.5,
            "default": 0
          },
          "currencyCode": {
            "type": "string",
            "description": "ISO 4217 currency code",
            "example": "USD",
            "default": "USD"
          },
          "color": {
            "type": "string",
            "description": "Hex color code for UI",
            "example": "#FF5733"
          },
          "icon": {
            "type": "string",
            "description": "Icon identifier for the app UI",
            "example": "credit-card"
          },
          "includeInTotal": {
            "type": "boolean",
            "description": "Include this account in the dashboard total balance",
            "default": true
          },
          "creditLimit": {
            "type": "number",
            "description": "Credit limit (only for CREDIT_CARD type)",
            "example": 5000
          },
          "statementClosingDay": {
            "type": "number",
            "description": "Statement closing day of month (1-31, only for CREDIT_CARD)",
            "example": 15
          },
          "paymentDueDay": {
            "type": "number",
            "description": "Payment due day of month (1-31, only for CREDIT_CARD)",
            "example": 5
          },
          "sortOrder": {
            "type": "number",
            "description": "Display order (lower = first)",
            "example": 0,
            "default": 0
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "AccountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm3xk7z9w0001jn08abc12345"
          },
          "name": {
            "type": "string",
            "example": "My Checking Account"
          },
          "type": {
            "type": "string",
            "enum": [
              "CASH",
              "DEBIT_CARD",
              "CREDIT_CARD",
              "SAVINGS",
              "INVESTMENT",
              "DIGITAL_WALLET",
              "OTHER"
            ],
            "example": "DEBIT_CARD"
          },
          "balance": {
            "type": "string",
            "example": "1500.5000",
            "description": "Current balance as string for precision"
          },
          "currencyCode": {
            "type": "string",
            "example": "USD"
          },
          "color": {
            "type": "string",
            "example": "#FF5733"
          },
          "icon": {
            "type": "string",
            "example": "credit-card"
          },
          "includeInTotal": {
            "type": "boolean",
            "example": true
          },
          "isArchived": {
            "type": "boolean",
            "example": false
          },
          "creditLimit": {
            "type": "string",
            "example": "5000.0000"
          },
          "statementClosingDay": {
            "type": "number",
            "example": 15
          },
          "paymentDueDay": {
            "type": "number",
            "example": 5
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          },
          "createdByUserId": {
            "type": "string",
            "example": "cm3xk7z9w0001jn08user12345"
          },
          "updatedByUserId": {
            "type": "string",
            "example": "cm3xk7z9w0001jn08user12345"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "balance",
          "currencyCode",
          "includeInTotal",
          "isArchived",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "AccountListResponseDto": {
        "type": "object",
        "properties": {
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 3,
            "description": "Total number of active (non-archived) accounts"
          }
        },
        "required": [
          "accounts",
          "total"
        ]
      },
      "AccountBalanceSummaryDto": {
        "type": "object",
        "properties": {
          "currencyCode": {
            "type": "string",
            "example": "USD"
          },
          "totalBalance": {
            "type": "string",
            "example": "5500.0000",
            "description": "Sum of all account balances in this currency"
          },
          "accountCount": {
            "type": "number",
            "example": 2,
            "description": "Number of accounts in this currency"
          }
        },
        "required": [
          "currencyCode",
          "totalBalance",
          "accountCount"
        ]
      },
      "AccountSummaryResponseDto": {
        "type": "object",
        "properties": {
          "balances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountBalanceSummaryDto"
            }
          },
          "totalAccounts": {
            "type": "number",
            "example": 5,
            "description": "Total number of active accounts"
          }
        },
        "required": [
          "balances",
          "totalAccounts"
        ]
      },
      "UpdateAccountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Account name",
            "example": "My Checking Account",
            "maxLength": 100
          },
          "balance": {
            "type": "number",
            "description": "Initial balance (defaults to 0)",
            "example": 1500.5,
            "default": 0
          },
          "color": {
            "type": "string",
            "description": "Hex color code for UI",
            "example": "#FF5733"
          },
          "icon": {
            "type": "string",
            "description": "Icon identifier for the app UI",
            "example": "credit-card"
          },
          "includeInTotal": {
            "type": "boolean",
            "description": "Include this account in the dashboard total balance",
            "default": true
          },
          "creditLimit": {
            "type": "number",
            "description": "Credit limit (only for CREDIT_CARD type)",
            "example": 5000
          },
          "statementClosingDay": {
            "type": "number",
            "description": "Statement closing day of month (1-31, only for CREDIT_CARD)",
            "example": 15
          },
          "paymentDueDay": {
            "type": "number",
            "description": "Payment due day of month (1-31, only for CREDIT_CARD)",
            "example": 5
          },
          "sortOrder": {
            "type": "number",
            "description": "Display order (lower = first)",
            "example": 0,
            "default": 0
          }
        }
      },
...
    }
  }
}